pyproject.toml
[tool.poetry]
name = "tinybar"
version = "0.3.2"
description = "Tiny MacOS menu bar utility to track price dynamics for assets on TinyMan"
authors = ["aorumbayev <aorumbayev at pm.me>"]

[tool.poetry.dependencies]
python = "^3.10"
rumps = "^0.4.0"
requests = "^2.26.0"
tinyman-py-sdk = {git = "https://github.com/tinymanorg/tinyman-py-sdk.git", rev = "main"}
py-algorand-sdk = "^1.8.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
py2app = "^0.28.0"
pytest = "^7.0.0"
black = "^21.12b0"
isort = "^5.10.1"
autoflake8 = "^0.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
